UAR-629: Refactoring with implementing the ARNO in the Base class for  ProtocolAttachmentProtocol instead of subclassing the irb child for the same class
