+++ src/main/java/edu/arizona/kra/irb/CustomProtocolForm.java
@@ -17,10 +17,8 @@ package edu.arizona.kra.irb;
 
 import org.kuali.kra.protocol.ProtocolFormBase;
 import org.kuali.kra.protocol.actions.ActionHelperBase;
-import org.kuali.kra.protocol.noteattachment.NotesAttachmentsHelperBase;
 
 import edu.arizona.kra.irb.actions.CustomActionHelper;
-import edu.arizona.kra.irb.noteattachment.CustomNotesAttachmentsHelper;
 
 /**
  * This is the custom code for the protocol form class for UofA
@@ -36,10 +34,6 @@ public class CustomProtocolForm extends org.kuali.kra.irb.ProtocolForm {
     protected ActionHelperBase createNewActionHelperInstanceHook(ProtocolFormBase protocolForm) throws Exception {
         return new CustomActionHelper((CustomProtocolForm)protocolForm);
     }
-    
-    @Override
-    protected NotesAttachmentsHelperBase createNewNotesAttachmentsHelperInstanceHook(ProtocolFormBase protocolForm) {
-        return new CustomNotesAttachmentsHelper(protocolForm);
-    }
+ 
 
 }
