--- src/main/java/edu/arizona/kra/irb/noteattachment/CustomProtocolAttachmentServiceImpl.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2005-2014 The Kuali Foundation
- * 
- * Licensed under the Educational Community License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- * http://www.opensource.org/licenses/ecl1.php
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package edu.arizona.kra.irb.noteattachment;
-
-import org.kuali.kra.irb.ProtocolDao;
-import org.kuali.kra.irb.noteattachment.ProtocolAttachmentServiceImpl;
-import org.kuali.kra.protocol.noteattachment.ProtocolAttachmentProtocolBase;
-import org.kuali.kra.protocol.noteattachment.ProtocolAttachmentService;
-import org.kuali.rice.krad.service.BusinessObjectService;
-
-
-/** UofA Customized implementation of {@link ProtocolAttachmentService ProtocolNoteAndAttachmentService}. */
-public class CustomProtocolAttachmentServiceImpl extends ProtocolAttachmentServiceImpl implements ProtocolAttachmentService {
-
-    public CustomProtocolAttachmentServiceImpl(BusinessObjectService boService, ProtocolDao protocolDao) {
-        super(boService, protocolDao);
-    }
-
-    @Override
-    public Class<? extends ProtocolAttachmentProtocolBase> getProtocolAttachmentProtocolClassHook() {
-        return CustomProtocolAttachmentProtocol.class;
-    }
-
-}
