--- src/main/resources/edu/arizona/kra/datadictionary/CustomProtocolAttachmentProtocol.xml
@@ -1,26 +1,0 @@
-<!--
-    Copyright 2005-2014 The Kuali Foundation
-    
-    Licensed under the Educational Community License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-    
-    http://www.osedu.org/licenses/ECL-2.0
-    
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
--->
-<beans xmlns="http://www.springframework.org/schema/beans"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xmlns:p="http://www.springframework.org/schema/p"
-    xsi:schemaLocation="http://www.springframework.org/schema/beans
-    http://www.springframework.org/schema/beans/spring-beans-2.0.xsd">
-
-    <bean id="ProtocolAttachmentProtocol" parent="ProtocolAttachmentProtocol-parentBean" >
-        <property name="businessObjectClass" value="edu.arizona.kra.irb.noteattachment.CustomProtocolAttachmentProtocol" />
-    </bean>
-
-</beans>
