+++ src/main/resources/org/kuali/kra/irb/IrbSpringBeans.xml
@@ -122,7 +122,7 @@
         </property>        
     </bean> 
         
-    <bean id="protocolAttachmentService" class="edu.arizona.kra.irb.noteattachment.CustomProtocolAttachmentServiceImpl">
+    <bean id="protocolAttachmentService" class="org.kuali.kra.irb.noteattachment.ProtocolAttachmentServiceImpl">
         <constructor-arg index="0" ref="businessObjectService"/>
         <constructor-arg index="1" ref="protocolDao"/>
         <property name = "personService" ref = "personService"/>
